Reinstate product groups (#3208)
Testing with full-sized GEFS found that the sheer number of tasks
overloads rocoto, resulting in `rocotorun` taking over 10 min to
complete or hanging entirely. To reduce the number of tasks, product
groups are reimplemented so that multiple forecast hour are processed in
a single task. However, the implementation is a little different than
previously.

The jobs where groups are enabled (atmos_products, oceanice_products,
wavepostsbs, atmos_ensstat, and gempak) have a new variable,
`MAX_TASKS`, that controls how many groups to use. This setting is
currently *per member*. The forecast hours to be processed are then
divided into this many groups as evenly as possible without crossing
forecast segment boundaries. The walltime for those jobs is then
multiplied by the number of times in the largest group. For the gridded
wave post job, the dependencies were also updated to trigger off of
either the data being available or the appropriate segment completing
(the dependencies had not been updated when the job was initially broken
into fhrs).

A number of helper methods are added to Tasks to determine these groups
and make a standard metatask variable dict in a centralized location.
There is also a function to multiply the walltime, but this may be
better off relocated to wxflow with the other time functions.

As part of switching from a single value to a list, hours are no longer
passed by rocoto as zero-padded values. The lists are comma-delimited
(without spaces) and split apart in the job stub (`jobs/rocoto/*`), so
each j-job call is still a single forecast hour.

The offline post (upp) job is not broken into groups, since it really
isn't used outside the analysis anymore.

Resolves #2999
Resolves #3210